General UI improvments and WIP for spellchecker suggestions.
Improved ListFragment usages by specifying an empty text and emptying it
when there is no data available. This is a better approach than to
finish the current activity.

Removes Toast usages where the information is providen by an empty
ListFragment.

The query hint on the SearchFragment SearchView updates with the user
checks different radio buttons. This way, there is no need for a
specific label to figure out what it does.

Improves the appearance of the RadioGroup by aligning it to the text
input in the SearchView.

Work in progress for the spelling checker provider so that we can get
live spelling suggestions from the getSpellingSuggestions API.

Specify compatibility for Java 1.7.